+++ package.json
@@ -30,6 +30,7 @@
         "jest-emotion": "^10.0.32",
         "lint-staged": "^10.5.1",
         "prop-types": "^15.7.2",
+        "random-username-generator": "^1.0.4",
         "react": "^16.12.0",
         "react-avatar-editor": "11.0.9",
         "react-dom": "^16.12.0",
